NETVIRT-1630 migrate to md-sal APIs
handles migration to new mdsal APIs for
    - databroker
    - Listener base classes
    - Readwrite txns

The following changes are the major in this review

1) import org.opendaylight.controller.md.sal.binding.api.DataBroker;
	to
   import org.opendaylight.mdsal.binding.api.DataBroker; (included blueprint)
2) .toJavaUtil(); to be removed
3) import com.google.common.base.Optional
       to
   import java.util.Optional
4) Optional.fromNullable to Optional.ofNullable
5) Optional.absent() to Optional.empty();
6) import
org.opendaylight.controller.md.sal.common.api.data.LogicalDatastoreType;
    to
import org.opendaylight.mdsal.common.api.LogicalDatastoreType;
7) import
org.opendaylight.controller.md.sal.binding.api.ReadOnlyTransaction;
    to
  import org.opendaylight.mdsal.binding.api.ReadTransaction;

8) ReadOnlyTransaction readTx = dataBroker.newReadOnlyTransaction()
    to
  ReadTransaction readTx = dataBroker.newReadOnlyTransaction()

9) writeTransaction/rw.submit().get() to writeTransaction.commit().get()

10) ReadFailedException to InterruptedException | ExecutionException

Dependency:
=============
This review is dependent on below GENIUS review

https://git.opendaylight.org/gerrit/c/genius/+/89001

+++ elanmanager/impl/src/main/java/org/opendaylight/netvirt/elan/l2gw/recovery/impl/L2GatewayConnectionInstanceRecoveryHandler.java
@@ -7,13 +7,14 @@
  */
 package org.opendaylight.netvirt.elan.l2gw.recovery.impl;
 
-import com.google.common.base.Optional;
+import java.util.Optional;
 import java.util.concurrent.ExecutionException;
 import javax.inject.Inject;
 import javax.inject.Singleton;
 
-import org.opendaylight.controller.md.sal.binding.api.DataBroker;
-import org.opendaylight.controller.md.sal.common.api.data.LogicalDatastoreType;
+import org.opendaylight.genius.datastoreutils.SingleTransactionDataBroker;
+import org.opendaylight.mdsal.binding.api.DataBroker;
+import org.opendaylight.mdsal.common.api.LogicalDatastoreType;
 import org.opendaylight.genius.infra.ManagedNewTransactionRunner;
 import org.opendaylight.genius.infra.ManagedNewTransactionRunnerImpl;
 import org.opendaylight.genius.mdsalutil.MDSALUtil;
@@ -71,22 +72,29 @@ public class L2GatewayConnectionInstanceRecoveryHandler implements ServiceRecove
                 .child(L2gatewayConnections.class)
                 .child(L2gatewayConnection.class, new L2gatewayConnectionKey(uuid));
 
-        Optional<L2gatewayConnection> l2gatewayConnectionOptional = MDSALUtil
-                .read(dataBroker, LogicalDatastoreType.CONFIGURATION, connectionInstanceIdentifier);
+        Optional<L2gatewayConnection> l2gatewayConnectionOptional = Optional.empty();
+        try {
+            l2gatewayConnectionOptional = SingleTransactionDataBroker.syncReadOptional(
+                    dataBroker, LogicalDatastoreType.CONFIGURATION, connectionInstanceIdentifier);
+        } catch (ExecutionException | InterruptedException e) {
+            LOG.error("recoverService: Exception while reading L2gatewayConnection DS", e);
+        }
         //l2GatewayConnectionUtils.addL2GatewayConnection(l2gatewayConnectionOptional.get());
 
-        L2gatewayConnection l2gatewayConnection = l2gatewayConnectionOptional.get();
+        if (l2gatewayConnectionOptional.isPresent()) {
+            L2gatewayConnection l2gatewayConnection = l2gatewayConnectionOptional.get();
 
-        try {
-            LOG.info("deleting l2 gateway connection {}",l2gatewayConnection.key());
-            txRunner.callWithNewWriteOnlyTransactionAndSubmit(
-                tx -> tx.delete(LogicalDatastoreType.CONFIGURATION, connectionInstanceIdentifier)).get();
-            LOG.info("recreating l2 gateway connection {}, {}",entityId, l2gatewayConnection.key());
-            txRunner.callWithNewWriteOnlyTransactionAndSubmit(
-                tx -> tx.put(LogicalDatastoreType.CONFIGURATION, connectionInstanceIdentifier,
-                        l2gatewayConnection)).get();
-        } catch (InterruptedException | ExecutionException e) {
-            LOG.error("Service recovery failed for l2gw connection {}", entityId);
+            try {
+                LOG.info("deleting l2 gateway connection {}",l2gatewayConnection.key());
+                txRunner.callWithNewWriteOnlyTransactionAndSubmit(
+                    tx -> tx.delete(LogicalDatastoreType.CONFIGURATION, connectionInstanceIdentifier)).get();
+                LOG.info("recreating l2 gateway connection {}, {}",entityId, l2gatewayConnection.key());
+                txRunner.callWithNewWriteOnlyTransactionAndSubmit(
+                    tx -> tx.put(LogicalDatastoreType.CONFIGURATION, connectionInstanceIdentifier,
+                            l2gatewayConnection)).get();
+            } catch (InterruptedException | ExecutionException e) {
+                LOG.error("Service recovery failed for l2gw connection {}", entityId);
+            }
         }
     }
